Deduplicate common JSON code
